Fix map state update
